backend/requirements.txt
# FastAPI and Server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
sse-starlette==2.1.3

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# AI & LangChain
langchain==0.3.9
langchain-openai==0.2.9
langchain-google-genai==2.0.5
langsmith<0.2.0,>=0.1.17
openai==1.57.2
google-generativeai==0.8.3
tiktoken==0.8.0

# Document Processing
pypdf2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
pandas==2.2.3
markdown==3.7

# Integrations
jira==3.8.0
boto3==1.35.73
httpx==0.27.2

# Task Queue (optional)
celery==5.4.0
redis==5.2.1

# Utilities
structlog==24.4.0
tenacity==9.0.0

